Handle stdin selection for editor execution
- Implemented logic to distinguish and set appropriate
  stdin (os.Stdin or /dev/tty) when launching the editor
- Ensures correct behavior whether standard input is a terminal or a pipe
